+++ src/snowflake/cli/plugins/spcs/image_registry/commands.py
@@ -1,8 +1,6 @@
-from click import ClickException
 from snowflake.cli.api.commands.snow_typer import SnowTyper
 from snowflake.cli.api.output.types import MessageResult, ObjectResult
 from snowflake.cli.plugins.spcs.image_registry.manager import (
-    NoImageRepositoriesFoundError,
     RegistryManager,
 )
 
@@ -27,9 +25,10 @@ def url(**options) -> MessageResult:
     Gets the image registry URL for the current account. Must be called from a
     role that can view at least one image repository in the image registry.
     """
-    try:
-        return MessageResult(RegistryManager().get_registry_url())
-    except NoImageRepositoriesFoundError:
-        raise ClickException(
-            "No image repository found. To get the registry url, please switch to a role with read access to at least one image repository or create a new image repository first."
-        )
+    return MessageResult(RegistryManager().get_registry_url())
+
+
+@app.command(requires_connection=True)
+def login(**options) -> MessageResult:
+    """Logs in to the account image registry with the current user's credentials. Must be called from a role that can view at least one image repository in the image registry."""
+    return MessageResult(RegistryManager().docker_registry_login().strip())